refactor: 누락된 메인 페이지 관련 슬라이스 병합 및 순환참조 해결 (#ATR-387)
