--- apps/service/entities/recent-article-item/model/index.ts
@@ -1,1 +1,0 @@
-export * from './type'
