--- apps/service/entities/recent-article-item/model/type/index.ts
@@ -1,1 +1,0 @@
-export * from './article'
