--- apps/service/entities/recent-article-item/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as RecentArticleItem } from './RecentArticleItem'
