--- apps/service/features/recent-articles/api/index.ts
@@ -1,1 +1,0 @@
-export { default as fetchArticles } from './fetchArticles'
