--- apps/service/features/recent-articles/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as RecentArticles } from './RecentArticles'
