+++ apps/service/widgets/main/api/index.ts
@@ -1,1 +1,2 @@
 export { default as getUserRanking } from './getUserRanking'
+export { default as fetchArticles } from './fetchArticles'
