+++ apps/service/widgets/main/model/index.ts
@@ -1,1 +1,2 @@
 export * from './type'
+export * from './hook'
