+++ apps/service/widgets/main/model/type/article.ts
