+++ apps/service/widgets/main/model/type/index.ts
@@ -1,1 +1,2 @@
 export * from './rankingType'
+export * from './article'
